+++ bin/nanocl/src/main.rs
@@ -126,28 +126,31 @@ async fn main() -> std::io::Result<()> {
 
 #[cfg(test)]
 mod tests {
-  use super::*;
+  use std::{path::Path, env};
 
-  use nanocld_client::NanocldClient;
+  use crate::utils::test::get_test_client;
+
+  use super::*;
 
   /// Test version command
   #[ntex::test]
   async fn version() {
-    assert_cli_ok!("nanocl", "version");
+    assert_cli_ok!("version");
   }
 
   /// Test Namespace commands
   #[ntex::test]
   async fn namespace() {
     const NAMESPACE_NAME: &str = "cli-namespace";
+
     // Try to create namespace
-    assert_cli_ok!("nanocl", "namespace", "create", NAMESPACE_NAME);
+    assert_cli_ok!("namespace", "create", NAMESPACE_NAME);
     // Try to list namespaces
-    assert_cli_ok!("nanocl", "namespace", "ls");
+    assert_cli_ok!("namespace", "ls");
     // Try to inspect namespace
-    assert_cli_ok!("nanocl", "namespace", "inspect", NAMESPACE_NAME);
+    assert_cli_ok!("namespace", "inspect", NAMESPACE_NAME);
     // Try to remove namespace
-    assert_cli_ok!("nanocl", "namespace", "rm", "-y", NAMESPACE_NAME);
+    assert_cli_ok!("namespace", "rm", "-y", NAMESPACE_NAME);
   }
 
   /// Test Cargo image commands
@@ -155,16 +158,15 @@ mod tests {
   async fn cargo_image() {
     const IMAGE_NAME: &str = "busybox:1.26.0";
     // Try to create cargo image
-    assert_cli_ok!("nanocl", "cargo", "image", "pull", IMAGE_NAME);
+    assert_cli_ok!("cargo", "image", "pull", IMAGE_NAME);
     // Try to list cargo images
-    assert_cli_ok!("nanocl", "cargo", "image", "ls");
+    assert_cli_ok!("cargo", "image", "ls");
     // Try to inspect cargo image
-    assert_cli_ok!("nanocl", "cargo", "image", "inspect", IMAGE_NAME);
+    assert_cli_ok!("cargo", "image", "inspect", IMAGE_NAME);
     // Try to remove cargo image
-    assert_cli_ok!("nanocl", "cargo", "image", "rm", "-y", IMAGE_NAME);
+    assert_cli_ok!("cargo", "image", "rm", "-y", IMAGE_NAME);
 
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "image",
       "import",
@@ -178,23 +180,30 @@ mod tests {
   async fn cargo() {
     const CARGO_NAME: &str = "cli-test";
     const IMAGE_NAME: &str = "ghcr.io/nxthat/nanocl-get-started:latest";
+    const NAMESPACE_NAME: Option<&str> = None;
+    let client = get_test_client();
     // Try to create cargo
-    assert_cli_ok!("nanocl", "cargo", "create", CARGO_NAME, IMAGE_NAME);
+    assert_cli_ok!("cargo", "create", CARGO_NAME, IMAGE_NAME);
+    assert_cargo_state!(client, CARGO_NAME, NAMESPACE_NAME, "created");
+
     // Try to list cargoes
-    assert_cli_ok!("nanocl", "cargo", "ls");
+    assert_cli_ok!("cargo", "ls");
     // Try to start a cargo
-    assert_cli_ok!("nanocl", "cargo", "start", CARGO_NAME);
+    assert_cli_ok!("cargo", "start", CARGO_NAME);
+    assert_cargo_state!(client, CARGO_NAME, NAMESPACE_NAME, "running");
     // Try to inspect a cargo
-    assert_cli_ok!("nanocl", "cargo", "inspect", CARGO_NAME);
+    assert_cli_ok!("cargo", "inspect", CARGO_NAME);
+    // Try to inspect cargo json
+    assert_cli_ok!("cargo", "inspect", "--display", "toml", CARGO_NAME);
+    // Try to inspect cargo toml
+    assert_cli_ok!("cargo", "inspect", "--display", "json", CARGO_NAME);
     // Try to patch a cargo
     assert_cli_ok!(
-      "nanocl", "cargo", "patch", CARGO_NAME, "--image", IMAGE_NAME, "--env",
-      "TEST=1",
+      "cargo", "patch", CARGO_NAME, "--image", IMAGE_NAME, "--env", "TEST=1",
     );
 
-    assert_cli_ok!("nanocl", "cargo", "history", CARGO_NAME);
-    let client =
-      NanocldClient::connect_to("http://ndaemon.nanocl.internal:8585", None);
+    assert_cli_ok!("cargo", "history", CARGO_NAME);
+    let client = get_test_client();
     let history = client
       .list_history_cargo(CARGO_NAME, None)
       .await
@@ -203,25 +212,27 @@ mod tests {
       .unwrap()
       .clone();
 
-    assert_cli_ok!(
-      "nanocl",
-      "cargo",
-      "revert",
-      CARGO_NAME,
-      &history.key.to_string(),
-    );
+    assert_cli_ok!("cargo", "revert", CARGO_NAME, &history.key.to_string());
 
     // Try to stop a cargo
-    assert_cli_ok!("nanocl", "cargo", "stop", CARGO_NAME);
+    assert_cli_ok!("cargo", "stop", CARGO_NAME);
+    assert_cargo_state!(client, CARGO_NAME, NAMESPACE_NAME, "exited");
     // Try to remove cargo
-    assert_cli_ok!("nanocl", "cargo", "rm", "-y", CARGO_NAME);
+    assert_cli_ok!("cargo", "rm", "-y", CARGO_NAME);
+    assert_cargo_not_exists!(client, CARGO_NAME, NAMESPACE_NAME);
+    // Try to run cargo
+    assert_cli_ok!("cargo", "run", CARGO_NAME, IMAGE_NAME);
+    assert_cargo_state!(client, CARGO_NAME, NAMESPACE_NAME, "running");
+
+    // Try to remove cargo
+    assert_cli_ok!("cargo", "rm", "-yf", CARGO_NAME);
+    assert_cargo_not_exists!(client, CARGO_NAME, NAMESPACE_NAME);
   }
 
   /// Test Resource commands
   #[ntex::test]
   async fn resource() {
     assert_cli_ok!(
-      "nanocl",
       "state",
       "apply",
       "-ys",
@@ -229,7 +240,6 @@ mod tests {
     );
 
     assert_cli_ok!(
-      "nanocl",
       "state",
       "apply",
       "-ys",
@@ -237,9 +247,8 @@ mod tests {
     );
 
     // History
-    assert_cli_ok!("nanocl", "resource", "history", "deploy-example.com");
-    let client =
-      NanocldClient::connect_to("http://ndaemon.nanocl.internal:8585", None);
+    assert_cli_ok!("resource", "history", "deploy-example.com");
+    let client = get_test_client();
     let history = client
       .list_history_resource("deploy-example.com")
       .await
@@ -248,28 +257,20 @@ mod tests {
       .unwrap()
       .clone();
     assert_cli_ok!(
-      "nanocl",
       "resource",
       "revert",
       "deploy-example.com",
       &history.key.to_string(),
     );
     // Remove resource
-    assert_cli_ok!("nanocl", "resource", "rm", "-y", "deploy-example.com");
-    assert_cli_ok!(
-      "nanocl",
-      "state",
-      "rm",
-      "-ys",
-      "../../examples/deploy_example.yml",
-    );
+    assert_cli_ok!("resource", "rm", "-y", "deploy-example.com");
+    assert_cli_ok!("state", "rm", "-ys", "../../examples/deploy_example.yml");
   }
 
   /// Test cargo exec command
   #[ntex::test]
   async fn cargo_exec() {
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "--namespace",
       "system",
@@ -281,7 +282,6 @@ mod tests {
     );
 
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "--namespace",
       "system",
@@ -294,7 +294,6 @@ mod tests {
     );
 
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "--namespace",
       "system",
@@ -306,7 +305,6 @@ mod tests {
     );
 
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "--namespace",
       "system",
@@ -318,7 +316,6 @@ mod tests {
     );
 
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "--namespace",
       "system",
@@ -332,41 +329,243 @@ mod tests {
   }
 
   #[ntex::test]
-  async fn state() {
-    assert_cli_ok!(
-      "nanocl",
+  async fn state_apply_default_statefile_name() {
+    let tests_path = Path::new("../../tests")
+      .canonicalize()
+      .expect("Can't cannonicalize tests folder path");
+    env::set_current_dir(tests_path).expect("Can't move in tests folder");
+    assert_cli_ok!("state", "apply", "-y");
+
+    let tests_path = Path::new("./without_s_option")
+      .canonicalize()
+      .expect("Can't cannonicalize without_s_option folder path");
+    env::set_current_dir(tests_path)
+      .expect("Can't move in without_s_option folder");
+    assert_cli_ok!("state", "apply", "-y");
+
+    let tests_path = Path::new("../without_s_option_yml")
+      .canonicalize()
+      .expect("Can't cannonicalize without_s_option_yml folder path");
+    env::set_current_dir(tests_path)
+      .expect("Can't move in without_s_option_yml folder");
+    assert_cli_ok!("state", "apply", "-y");
+
+    let tests_path = Path::new("../../bin/nanocl")
+      .canonicalize()
+      .expect("Can't cannonicalize tests folder path");
+    env::set_current_dir(tests_path).expect("Can't move back in nanocl folder");
+
+    assert_cli_err!("state", "apply", "-y");
+  }
+
+  #[ntex::test]
+  async fn state_apply_invalid_image() {
+    assert_cli_err!(
       "state",
       "apply",
       "-ys",
-      "../../examples/deploy_example.yml",
+      "../../tests/invalid_init_container_image.yml",
+    );
+    assert_cli_err!(
+      "state",
+      "apply",
+      "-ys",
+      "../../tests/invalid_container_image.yml",
+    );
+  }
+
+  #[ntex::test]
+  async fn state_apply_url_statefile() {
+    assert_cli_ok!(
+      "state",
+      "apply",
+      "-ys",
+      "https://raw.githubusercontent.com/nxthat/nanocl/nightly/examples/deploy_example.yml",
     );
 
     assert_cli_ok!(
-      "nanocl",
+      "state",
+      "rm",
+      "-ys",
+      "https://raw.githubusercontent.com/nxthat/nanocl/nightly/examples/deploy_example.yml",
+    );
+
+    assert_cli_err!("state", "rm", "-ys", "https://google.com");
+
+    assert_cli_err!(
+      "state",
+      "rm",
+      "-ys",
+      "https://feiwfioewjnoifjnewoifnoiwef.fr",
+    );
+  }
+
+  #[ntex::test]
+  async fn state_apply_binds() {
+    assert_cli_ok!("state", "apply", "-ys", "../../tests/relative_bind.yml");
+    assert!(
+      Path::new("./toto")
+        .canonicalize()
+        .expect("Can't cannonicalize bind toto folder path")
+        .exists(),
+      "Relative bind was not created",
+    );
+    assert!(
+      Path::new("/tmp/toto")
+        .canonicalize()
+        .expect("Can't cannonicalize bind /tmp/toto folder path")
+        .exists(),
+      "Relative bind was not created",
+    );
+    assert_cli_ok!("state", "rm", "-ys", "../../tests/relative_bind.yml");
+  }
+
+  #[ntex::test]
+  async fn state_apply_invalid_statefile() {
+    assert_cli_err!("state", "apply", "-ys", "../../tests/invalid_yaml.yaml");
+    assert_cli_err!(
+      "state",
+      "apply",
+      "-ys",
+      "../../examples/invalid_json.json",
+    );
+
+    assert_cli_err!(
+      "state",
+      "apply",
+      "-ys",
+      "../../examples/invalid_toml.toml",
+    );
+    assert_cli_err!(
+      "state",
+      "apply",
+      "-ys",
+      "../../examples/invalid_statefile.yaml",
+    );
+    assert_cli_err!(
+      "state",
+      "apply",
+      "-ys",
+      "../../examples/invalid_statefile.toml",
+    );
+  }
+
+  #[ntex::test]
+  async fn state_apply_toml() {
+    let client = get_test_client();
+    const DEPLOY_CARGO_NAME: &str = "deploy-example";
+    const DEPLOY_CARGO2_NAME: &str = "deploy-example2";
+    const DEPLOY_NAMESPACE_NAME: Option<&str> = None;
+
+    assert_cli_ok!(
       "state",
       "apply",
       "-ys",
       "../../examples/deploy_example.toml",
     );
-
-    assert_cli_ok!(
-      "nanocl",
-      "state",
-      "apply",
-      "-rys",
-      "../../examples/deploy_example.toml"
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO2_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
     );
 
+    assert_cli_ok!("state", "rm", "-ys", "../../examples/deploy_example.toml");
+    assert_cargo_not_exists!(client, DEPLOY_CARGO_NAME, DEPLOY_NAMESPACE_NAME);
+    assert_cargo_not_exists!(client, DEPLOY_CARGO2_NAME, DEPLOY_NAMESPACE_NAME);
+  }
+
+  #[ntex::test]
+  async fn state_apply_json() {
+    let client = get_test_client();
+    const DEPLOY_CARGO_NAME: &str = "deploy-example";
+    const DEPLOY_CARGO2_NAME: &str = "deploy-example2";
+    const DEPLOY_NAMESPACE_NAME: Option<&str> = None;
+
     assert_cli_ok!(
-      "nanocl",
+      "state",
+      "apply",
+      "-ys",
+      "../../examples/deploy_example.json",
+    );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO2_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
+
+    assert_cli_ok!("state", "rm", "-ys", "../../examples/deploy_example.json");
+
+    assert_cargo_not_exists!(client, DEPLOY_CARGO_NAME, DEPLOY_NAMESPACE_NAME);
+    assert_cargo_not_exists!(client, DEPLOY_CARGO2_NAME, DEPLOY_NAMESPACE_NAME);
+  }
+
+  #[ntex::test]
+  async fn state_logs_invalide_statefile_kind() {
+    assert_cli_err!("state", "logs", "-s", "../../examples/secret_env.yml");
+  }
+
+  #[ntex::test]
+  async fn state() {
+    let client = get_test_client();
+    const DEPLOY_CARGO_NAME: &str = "deploy-example";
+    const DEPLOY_CARGO2_NAME: &str = "deploy-example2";
+    const DEPLOY_NAMESPACE_NAME: Option<&str> = None;
+    const CARGO_NAME: &str = "cargo-example";
+    const CARGO_NAMESPACE_NAME: Option<&str> = Some("cargo-example");
+
+    assert_cli_ok!(
       "state",
       "apply",
       "-pys",
       "../../examples/deploy_example.yml",
     );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO2_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
 
     assert_cli_ok!(
-      "nanocl",
+      "state",
+      "apply",
+      "-rys",
+      "../../examples/deploy_example.toml"
+    );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
+    assert_cargo_state!(
+      client,
+      DEPLOY_CARGO2_NAME,
+      DEPLOY_NAMESPACE_NAME,
+      "running"
+    );
+
+    assert_cli_ok!(
       "state",
       "logs",
       "-t",
@@ -376,48 +575,34 @@ mod tests {
       "../../examples/deploy_example.yml",
     );
 
-    assert_cli_ok!(
-      "nanocl",
-      "state",
-      "apply",
-      "-ys",
-      "../../examples/cargo_example.yml",
-    );
+    assert_cli_ok!("state", "logs", "-s", "../../examples/deploy_example.yml");
 
-    assert_cli_ok!(
-      "nanocl",
-      "state",
-      "apply",
-      "-ys",
-      "../../examples/cargo_example.yml",
-    );
+    assert_cli_ok!("state", "rm", "-ys", "../../examples/deploy_example.yml");
+    assert_cargo_not_exists!(client, DEPLOY_CARGO_NAME, DEPLOY_NAMESPACE_NAME);
+    assert_cargo_not_exists!(client, DEPLOY_CARGO2_NAME, DEPLOY_NAMESPACE_NAME);
 
-    assert_cli_ok!(
-      "nanocl",
-      "state",
-      "rm",
-      "-ys",
-      "../../examples/cargo_example.yml",
-    );
+    assert_cli_ok!("state", "apply", "-ys", "../../examples/cargo_example.yml");
+    assert_cargo_state!(client, CARGO_NAME, CARGO_NAMESPACE_NAME, "running");
 
-    assert_cli_ok!(
-      "nanocl",
-      "state",
-      "rm",
-      "-ys",
-      "../../examples/deploy_example.yml",
-    );
+    assert_cli_ok!("state", "apply", "-ys", "../../examples/cargo_example.yml");
+    assert_cargo_state!(client, CARGO_NAME, CARGO_NAMESPACE_NAME, "running");
+
+    assert_cli_ok!("state", "rm", "-ys", "../../examples/cargo_example.yml");
+    assert_cargo_not_exists!(client, CARGO_NAME, CARGO_NAMESPACE_NAME);
   }
 
   #[ntex::test]
   async fn info() {
-    assert_cli_ok!("nanocl", "info");
+    assert_cli_ok!("info");
   }
 
   #[ntex::test]
   async fn cargo_run() {
+    let client = get_test_client();
+    const CARGO_NAME: &str = "cli-test-run";
+    const NAMESPACE_NAME: Option<&str> = None;
+
     assert_cli_ok!(
-      "nanocl",
       "cargo",
       "run",
       "cli-test-run",
@@ -425,9 +610,23 @@ mod tests {
       "-e",
       "MESSAGE=GREETING",
     );
+    assert_cargo_state!(client, CARGO_NAME, NAMESPACE_NAME, "running");
 
-    assert_cli_ok!("nanocl", "cargo", "stop", "cli-test-run");
-    assert_cli_ok!("nanocl", "cargo", "rm", "-y", "cli-test-run");
+    assert_cli_ok!("cargo", "stop", "cli-test-run");
+    assert_cargo_state!(client, CARGO_NAME, NAMESPACE_NAME, "exited");
+
+    assert_cli_ok!("cargo", "rm", "-y", "cli-test-run");
+    assert_cargo_not_exists!(client, CARGO_NAME, NAMESPACE_NAME);
+  }
+
+  #[ntex::test]
+  async fn cargo_inspect_invalid() {
+    assert_cli_err!("cargo", "inspect", "ewfwefew");
+  }
+
+  #[ntex::test]
+  async fn cargo_logs() {
+    assert_cli_ok!("cargo", "-n", "system", "logs", "nanocld");
   }
 
   // #[ntex::test]
@@ -475,6 +674,11 @@ mod tests {
 
   #[ntex::test]
   async fn node_list() {
-    assert_cli_ok!("nanocl", "node", "ls");
+    assert_cli_ok!("node", "ls");
+  }
+
+  #[ntex::test]
+  async fn http_logs() {
+    assert_cli_ok!("system", "http", "logs");
   }
 }